Workaround the RSS Feed update
The standard RSS update with XmlListModel/ListView does not work by some reason.
After some time, the ListView gets out of sync with the XMLListModel and it shows
wrong content (already deleted emails, nothing at all etc.). It maybe caused by
some incorrect "id" field in the RSS entries? I don't know, but if the "id" as key
removed from the model, the all items are always updated in the RSS fetch causing
"new email arrived" messages always
The current workaround contains two XmlListModels, one to calculate the correct
amount of new emails with "id" as key and a second without "id" associated with the
ListView. Since the second updates all items always, it is unlikely that the ListView
goes out-of-sync.